11-8일 로그인 + 회원가입 가능

+++ app/src/main/java/com/lx/api/BasicApi.kt
@@ -67,8 +67,8 @@ interface BasicApi {
     fun memberArea(
         @Field("requestCode") requestCode: String,
         @Field("id") id: String,
-        @Field("lat") lat: Double,
-        @Field("lng") lng: Double
+        @Field("lat") lat: Int,
+        @Field("lng") lng: Int
     ): Call<MemberAreaResponse>
 
     /**
